+++ docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.0.1 (Unreleased)
+------------------
+- Enh: Allow translation of non-enabled modules
+
+
 1.0.0 (June 10, 2020)
 -----------------------
 - Enh: 1.4 nonce support